requirements.txt
fastapi
uvicorn
openai>=1.0.0
azure-search-documents
azure-core
markdown2
jinja2
python-multipart
